Added item class
